Move test specific condition out of the api
